requirements.txt
asgiref==3.5.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
coverage==6.3.2
Django==4.0.3
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-js-asset==2.0.0
django-mptt==0.13.4
idna==3.3
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stripe==2.67.0
tomli==2.0.1
typing-extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
